Code refactored
Offers refactored to request
