+++ app/src/main/java/app/chaffer/adapter/LocationListAdapter.java
@@ -1,23 +1,15 @@
 package app.chaffer.adapter;
 
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.os.AsyncTask;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
 import android.widget.TextView;
 
-import java.io.InputStream;
 import java.util.List;
 
 import app.chaffer.Location;
-import app.chaffer.Offer;
 import app.chaffer.R;
-import de.hdodenhof.circleimageview.CircleImageView;
 
 /**
  * Created by Mac on 12/02/2018.
