+++ app/src/main/res/layout/fragment_home.xml
@@ -174,7 +174,7 @@
                     android:layout_height="wrap_content"
                     android:layout_centerVertical="true"
                     android:layout_weight="1"
-                    android:text="  Place Offer"
+                    android:text="  Place Request"
                     android:textStyle="bold" />
 
                 <Button
@@ -182,7 +182,7 @@
                     android:layout_width="wrap_content"
                     android:textStyle="bold"
                     android:layout_centerVertical="true"
-                    android:text="  View Offers"
+                    android:text="  View Request"
                     android:layout_weight="1"
                     android:layout_height="wrap_content" />
