+++ app/src/main/res/layout/fragment_offer_placement.xml
@@ -1,98 +1,102 @@
-<?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    android:background="@color/cardview_shadow_end_color"
+    android:padding="@dimen/standard_padding"
+    tools:context="app.chaffer.Fragments.FragmentOfferPlacement">
 
-    <include
-        android:layout_above="@+id/card_view_buttons"
+    <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        layout="@layout/activity_maps"/>
-
-    <android.support.v7.widget.CardView
-        android:layout_above="@id/card_view_buttons"
-        android:id="@+id/card_view_search_bar"
-        android:layout_width="match_parent"
-        android:layout_marginBottom="3dp"
-        app:cardCornerRadius="@dimen/cardview_corner_radius"
-        android:layout_margin="2dp"
-        android:layout_height="wrap_content">
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true">
 
 
-
-        <EditText
-
-            android:id="@+id/editText_enterLocation_text"
+        <android.support.v7.widget.CardView
+            android:id="@+id/card_view_description"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="Enter location"
-            android:background="@android:color/transparent"/>
+            android:layout_below="@id/card_view_spinner"
+            android:layout_marginTop="@dimen/margin_top"
+            app:cardBackgroundColor="#e7e6e6"
+            app:cardCornerRadius="@dimen/cardview_corner_radius">
 
 
+            <EditText
+                android:id="@+id/editText_description"
+                android:layout_width="match_parent"
+                android:layout_height="300px"
+                android:layout_alignParentLeft="true"
+                android:layout_below="@+id/text_description"
+                android:background="@android:color/transparent"
+                android:hint="Describe your request."
+                android:maxLines="5"
+                android:textColor="@color/colorAccent" />
+
+        </android.support.v7.widget.CardView>
 
 
-    </android.support.v7.widget.CardView>
-
-    <android.support.v7.widget.CardView
-        android:id="@+id/card_view_buttons"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentEnd="true"
-        android:layout_marginBottom="3dp"
-        android:layout_above="@+id/btn_next"
-        android:layout_alignParentRight="true">
-
-    <LinearLayout
-        android:id="@+id/l1"
-        android:layout_width="match_parent"
-        android:weightSum="3"
-        android:layout_height="wrap_content">
-
-    <Button
-        android:id="@+id/btn_pickup_point"
-        android:layout_weight="1" android:textColor="@color/white"
-        android:textStyle="bold"
-        android:text="Pickup point"
-        android:background="@color/colorPrimary"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        />
-
-        <Button
-            android:id="@+id/btn_remover_markers"
-            android:text="Remover markers"
-            android:layout_weight="1"
-            android:textColor="@color/colorPrimary"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content" />
-
-
-        <Button
-            android:id="@+id/btn_delivery_point"
-            android:layout_width="wrap_content"
-            android:layout_weight="1"
-            android:text="Delivery Point"
-            android:background="@color/colorPrimary"
+        <android.support.v7.widget.CardView
+            android:id="@+id/card_view_pickup_locationDescription"
+            android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:textColor="@color/white"
-            android:textStyle="bold"
-            />
+            android:layout_below="@id/card_view_description"
+            android:layout_marginTop="@dimen/margin_top"
+            app:cardBackgroundColor="@color/gray"
+            app:cardCornerRadius="@dimen/cardview_corner_radius">
 
-    </LinearLayout>
-    </android.support.v7.widget.CardView>
+
+            <EditText
+                android:id="@+id/editText_time"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_alignParentLeft="true"
+                android:layout_below="@+id/text_time"
+                android:background="@android:color/transparent"
+                android:hint="Time to deliver"
+                android:maxLines="5"
+                android:textColor="@color/colorAccent" />
+
+        </android.support.v7.widget.CardView>
+
+
+        <android.support.v7.widget.CardView
+            android:id="@+id/card_view_dropp_offlocationDescription"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/card_view_pickup_locationDescription"
+            android:layout_marginTop="@dimen/margin_top"
+            app:cardBackgroundColor="@color/gray"
+            app:cardCornerRadius="@dimen/cardview_corner_radius">
+
+
+            <EditText
+                android:id="@+id/editText_amount"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_alignParentLeft="true"
+                android:layout_below="@+id/text_amount"
+                android:background="@android:color/transparent"
+                android:hint="Amount"
+                android:maxLines="5"
+                android:textColor="@color/colorAccent" />
+
+        </android.support.v7.widget.CardView>
+    </RelativeLayout>
 
 
     <Button
-        android:id="@+id/btn_next"
+        android:id="@+id/place_offer"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:textColor="@color/white"
-        android:textStyle="bold"
-        android:background="@color/colorPrimary"
         android:layout_alignParentBottom="true"
-        android:text="Next" />
-
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:layout_marginTop="@dimen/margin_top"
+        android:background="@color/colorPrimary"
+        android:text="Place request"
+        android:textColor="@color/white" />
 
 
 </RelativeLayout>
