+++ app/src/main/res/layout/fragment_view_request_details.xml
